Added ``FacetField`` & changed the way facets are processed.
Facet data is no longer quietly duplicated just before it goes into the index. Instead, full fields are created (with all the standard data & methods) to contain the faceted information.

This change is backward-compatible, but allows for better extension, not requiring data duplication into an unfaceted field and a little less magic.