Mkae sure that the SearchIndex knows about the backend
This is a fix aimed at being a bugfix, not the solution to the problem.
The current API doesn't have a clean way to pass a site all the way
through the index to the backend unless you define what Site an index
belongs to when defining the index.

Personally, I would modify the `site.register` code to set this and/or
have the code that generates `index.backend` capable of looking at what
site it is a part of.

Both of those solutions tread a bit heavier than fixing hte issue of
`update_index` not working, however, so I've opted for the easier fix.