Fall back to active_window if view.window() is None
Fun discovery: if you switch tabs while a command is working,
self.view.window() is None. (Admittedly this is a consequence
of my deciding to do async command processing... but, hey,
got to live with that now.)
I did try tracking the window used at the start of the command
and using it instead of view.window() later, but that results
panels on a non-visible window, which is especially useless in
the case of the quick panel.
So, this is not necessarily ideal, but it does work.