facet_counts(): avoid duplicate query if the query has already executed
As discussed in #439, it's common for a SearchQuerySet to result in a completely
duplicate query when .facet_counts() is called after the query has already been
executed. In this common case, we simply return the facet results from that
query directly; if for some reason the user wanted to get facet counts before
running (or customizing) the main query this will behave identically to v1.2.5.