Added setting allowing 'Git: Status' to open files
One of the biggest issues I'd have with the Git plugin
was that there was no easy way to get to files you've
been working on.  For instance, when resolving merge
conflicts, or picking up after switching projects.
Enabling this option lets you quickly open modified files,
conflicts and such.