blame: scroll scratch buffer
When running `git blame` with a full-file context, this change causes the
newly-created scratch buffer to automatically scroll to the same position
the user was in when he/she called `git blame`. This way, the user will be
looking at the blame lines that are relevant to what he/she was interested in
when calling `git blame`.

Note that if the user has word_wrap turned on, and the lines in the blame buffer
are longer (and wrap more) than the lines in the original buffer, then the
scrolled-to position in the blame buffer may be slightly above the position
in the original file. This is because the position is calculated by the offset
from the top of the file, not by the number of lines of text.