Add 'models' kwarg to ``run_mlt`` method for both ``SolrSearchQuery`` and ``ElasticsearchSearchQuery``.
Also, reverted changes made to ``solr_tests.LiveSolrMoreLikeThisTestCase.test_more_like_this`` made in SHA: ad79f05b that seem to break the tests for the ``more_like_this`` method. Also used ``sorted`` to order the result PKs and expected output; this was done in lieu of ``assertItemsEqual``. The failure of the test on this front was masking the breakage of the interaction between ``more_like_this`` and ``models`` methods.