app loading: fix support for multi-level app labels
The previous code assumed only one level of hierarchy in the
app label (e.g. “core.MyModel”) but Django is now documented
as only requiring that the *last* component before the model
be unique (e.g. “myapp.contrib.myapp_django.MyModel”) across
the entire project. In Django 1.6 this is a hard requirement
but was relaxed in Django 1.7 where AppConfig allows the app
label to be customized:

https://docs.djangoproject.com/en/1.6/ref/settings/#installed-apps
https://docs.djangoproject.com/en/1.7/ref/settings/#installed-apps