Fix concurrency issues when building UnifiedIndex
We were getting this error a lot when under load in a multithreaded wsgi
environment:

    Model '%s' has more than one 'SearchIndex`` handling it.

Turns out the connections in haystack.connections and the UnifiedIndex
instance were stored globally. However there is a race condition in
UnifiedIndex.build() when multiple threads both build() at once,
resulting in the above error.

Best fix is to never share the same engine or UnifiedIndex across
multiple threads. This commit does that.

Closes #959
Closes #615