SearchQuerySet._clone() to also clone using value
I came across an issue when using a different index.  Even though I had created my SearchQuerySet with using='something', the _clone method was creating a clone with using='default' -- which was causing a problem since I did not have a 'default' index.

I've updated the code to maintain the self._using when cloning.  I have fixed this in my own project by sub-classing SearchQuerySet and overriding the _clone method to be the same as this changeset.