Update Solr spelling suggestion handling
The support matrix for this is a problem since the Solr response format changes based on the version,
configuration, and query parameters (i.e. spellcheck.collateExtendedResults) so this is moved into a separate function which logs errors and honors
the backend fail silently setting.

This has been tested using Solr 6.4 and 6.5 with both
the regular and collateExtendedResults formats.