Fix UnicodeDecodeError in error message
Because of the way the default __repr__ works in Django models, we can get a
UnicodeDecodeError when creating the SearchFieldError if a model does not have
an attribute. eg:
UnicodeDecodeError: 'ascii' codec can't decode byte 0xc3 in position 11: ordinal not in range(128)
and this hides the real problem.

I have left alone the other SearchFieldError in this method because current_obj is always
None. The error message is a bit strange in this case but it won't suffer from the same problem.