Order queryset by pk in update batching This solves #1615
The queryset is not ordered by pk by default, however the batching filter relies on the results being ordered.
When the results are not ordered by pk, some objects are not indexed.
This can happen when the underlying database doesn't have default ordering by pk, or when the model or index_queryset() have a different ordering.