List elasticsearch as an optional dependency
Currently, if someone follows the tutorial using elasticsearch, they'll
end up with a broken setup that does not work. The error will indicate
that an additional, unlisted, dependency is required.

Upon installing `elasticsearch`, there's still some very unclear errors.
After digging through issues and the documentation, it turns out that a
**specific** version of the package is required.

This changeset lists `elasticsearch` as an extra ("optional")
dependency, and mentions it in the tutorial. This should make life a bit
easier for new users.