Fix SearchQuerySet to use all given words in db search
With this fix, when a string containing more than one word is passed to the SearchQuerySet filter, a database query will be created that uses all the specified words and combines them into a single WHERE clause. The returned result should be all records that contain all those specified words.