Use standard Solr ports
The old code assumed the non-standard port which the test runner has used for a long, long time to avoid conflicts with other Solr instances on the same host. Since that’s a) ancient history and b) moot in an isolated CI environment, we’ll drop it rather than remapping.